Backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
contourpy==1.1.1
cryptography==42.0.5
cycler==0.12.1
DateTime==5.4
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
fonttools==4.49.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.7.5
numpy==1.24.4
packaging==23.2
pandas==2.0.3
passlib==1.7.4
pillow==10.2.0
pyasn1==0.5.1
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
PyMySQL==1.1.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.1
zipp==3.17.0
zope.interface==6.2
